backend/requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
easyocr
streamlit
requests
pydantic
pydantic[email]
python-multipart
sentence-transformers
faiss-cpu
numpy
openai
sqlalchemy
databases
passlib[bcrypt]
python-dotenv
databases[postgresql]
asyncpg
aiosmtplib
pytest
pytest-asyncio
httpx
stripe
sentry-sdk
jwt
Flask
Flask-CORS
opencv-python 
speechrecognition 
moviepy==1.0.3 
transformers 
torch 
pillow 
numpy
pyPDF2
python-docx
python-jose[cryptography]
authlib
alembic
psycopg2-binary